Extracted setup code from sensor app. Created a custom "smart-environment-shared" module, that can hold all code, that is required by the sensor as well as the actor node.
