+++ sensor/LICENSE
