+++ sensor/endpoints.c
