+++ shared/Makefile
@@ -1,0 +1,5 @@
+MODULE=smart-environment-shared
+
+RIOTBASE ?= $(CURDIR)/../../RIOT
+
+include $(RIOTBASE)/Makefile.base
