+++ shared/setup.h
